+++ src/main/java/org/fedoraproject/javapackages/validator/util/BytecodeVersionJarValidator.java
@@ -1,4 +1,4 @@
-package org.fedoraproject.javapackages.validator.helpers;
+package org.fedoraproject.javapackages.validator.util;
 
 import java.io.ByteArrayInputStream;
 import java.nio.ByteBuffer;
@@ -10,7 +10,6 @@ import java.util.jar.JarEntry;
 import java.util.jar.JarInputStream;
 
 import org.apache.commons.compress.archivers.cpio.CpioArchiveEntry;
-import org.fedoraproject.javapackages.validator.Common;
 import org.fedoraproject.javapackages.validator.spi.Decorated;
 import org.fedoraproject.javapackages.validator.spi.TestResult;
