+++ src/main/java/org/fedoraproject/javapackages/validator/util/Common.java
@@ -1,8 +1,6 @@
-package org.fedoraproject.javapackages.validator;
+package org.fedoraproject.javapackages.validator.util;
 
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.PrintStream;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -10,8 +8,6 @@ import java.util.SortedMap;
 import java.util.TreeMap;
 
 import org.apache.commons.compress.archivers.cpio.CpioArchiveEntry;
-import org.fedoraproject.javapackages.validator.spi.Decorated;
-import org.fedoraproject.javapackages.validator.spi.LogEntry;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.kojan.javadeptools.rpm.RpmArchiveInputStream;
@@ -70,13 +66,5 @@ public class Common {
         }
 
         return result;
-    }
-
-    public static LogEntry logException(Exception ex) {
-        var stackTrace = new ByteArrayOutputStream();
-        ex.printStackTrace(new PrintStream(stackTrace, false, StandardCharsets.UTF_8));
-        return LogEntry.error("An exception occured:{0}{1}",
-                Decorated.plain(System.lineSeparator()),
-                Decorated.plain(new String(stackTrace.toByteArray(), StandardCharsets.UTF_8)));
     }
 }
