+++ src/main/java/org/fedoraproject/javapackages/validator/util/DuplicateFileValidator.java
@@ -1,4 +1,4 @@
-package org.fedoraproject.javapackages.validator.helpers;
+package org.fedoraproject.javapackages.validator.util;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -10,8 +10,6 @@ import java.util.TreeMap;
 
 import org.apache.commons.compress.archivers.cpio.CpioArchiveEntry;
 import org.apache.commons.lang3.tuple.Pair;
-import org.fedoraproject.javapackages.validator.Common;
-import org.fedoraproject.javapackages.validator.DefaultValidator;
 import org.fedoraproject.javapackages.validator.spi.Decorated;
 
 import io.kojan.javadeptools.rpm.RpmInfo;
