+++ src/main/java/org/fedoraproject/javapackages/validator/validators/AttributeBuildRequiresValidator.java
@@ -1,6 +1,6 @@
 package org.fedoraproject.javapackages.validator.validators;
 
-import org.fedoraproject.javapackages.validator.helpers.ElementwiseValidator;
+import org.fedoraproject.javapackages.validator.util.ElementwiseValidator;
 
 import io.kojan.javadeptools.rpm.RpmInfo;
 import io.kojan.javadeptools.rpm.RpmPackage;
