+++ src/main/java/org/fedoraproject/javapackages/validator/validators/BytecodeVersionValidator.java
@@ -3,8 +3,8 @@ package org.fedoraproject.javapackages.validator.validators;
 import java.nio.file.Path;
 import java.util.Map;
 
-import org.fedoraproject.javapackages.validator.helpers.BytecodeVersionJarValidator;
 import org.fedoraproject.javapackages.validator.spi.Decorated;
+import org.fedoraproject.javapackages.validator.util.BytecodeVersionJarValidator;
 
 import io.kojan.javadeptools.rpm.RpmPackage;
