+++ src/main/java/org/fedoraproject/javapackages/validator/validators/JavadocNoarchValidator.java
@@ -1,8 +1,8 @@
 package org.fedoraproject.javapackages.validator.validators;
 
-import org.fedoraproject.javapackages.validator.Common;
-import org.fedoraproject.javapackages.validator.helpers.ElementwiseValidator;
 import org.fedoraproject.javapackages.validator.spi.Decorated;
+import org.fedoraproject.javapackages.validator.util.Common;
+import org.fedoraproject.javapackages.validator.util.ElementwiseValidator;
 
 import io.kojan.javadeptools.rpm.RpmPackage;
