+++ src/main/java/org/fedoraproject/javapackages/validator/validators/NVRJarMetadataValidator.java
@@ -9,11 +9,11 @@ import java.util.TreeMap;
 import java.util.jar.JarInputStream;
 
 import org.apache.commons.compress.archivers.cpio.CpioArchiveEntry;
-import org.fedoraproject.javapackages.validator.Common;
-import org.fedoraproject.javapackages.validator.DefaultValidator;
-import org.fedoraproject.javapackages.validator.helpers.JarValidator;
-import org.fedoraproject.javapackages.validator.helpers.RpmJarConsumer;
 import org.fedoraproject.javapackages.validator.spi.Decorated;
+import org.fedoraproject.javapackages.validator.util.Common;
+import org.fedoraproject.javapackages.validator.util.DefaultValidator;
+import org.fedoraproject.javapackages.validator.util.JarValidator;
+import org.fedoraproject.javapackages.validator.util.RpmJarConsumer;
 
 import io.kojan.javadeptools.rpm.RpmPackage;
