+++ src/main/java/org/fedoraproject/javapackages/validator/validators/NoBootstrapValidator.java
@@ -1,7 +1,7 @@
 package org.fedoraproject.javapackages.validator.validators;
 
-import org.fedoraproject.javapackages.validator.helpers.ElementwiseValidator;
 import org.fedoraproject.javapackages.validator.spi.Decorated;
+import org.fedoraproject.javapackages.validator.util.ElementwiseValidator;
 
 import io.kojan.javadeptools.rpm.RpmPackage;
