+++ src/main/java/org/fedoraproject/javapackages/validator/validators/jp
@@ -1,0 +1,1 @@
+/home/mkoncek/Work/Fedora-tests/javapackages/src
