+++ src/test/java/org/fedoraproject/javapackages/validator/validators/RpmFilesizeValidatorTest.java
@@ -7,7 +7,7 @@ import java.nio.file.Paths;
 
 import org.easymock.EasyMock;
 import org.fedoraproject.javapackages.validator.TestCommon;
-import org.fedoraproject.javapackages.validator.helpers.RpmFilesizeValidator;
+import org.fedoraproject.javapackages.validator.util.RpmFilesizeValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
